Add accounts-tool subcommand under ledger-tool
